+++ Runtime/Internal/StateFactoryData.cs
@@ -7,6 +7,6 @@ namespace GameLovers.Statechart.Internal
 	internal struct StateFactoryData
 	{
 		public Action<IStatechartEvent> StateChartMoveNextCall;
-		public IStatechartDebug Statechart;
+		public IStateMachineDebug StateMachine;
 	}
 }