+++ Runtime/StatechartEvent.cs
@@ -6,7 +6,7 @@ namespace GameLovers.Statechart
 {
 	/// <summary>
 	/// Events are unique inputs to make the State Chart move forward when defined in the setup.
-	/// It requires to call <see cref="IStatechart.Trigger(IStatechartEvent)"/> to move the State Chart forward.
+	/// It requires to call <see cref="IStateMachine.Trigger(IStatechartEvent)"/> to move the State Chart forward.
 	/// </summary>
 	public interface IStatechartEvent : IEquatable<IStatechartEvent>
 	{