+++ Tests/Editor/StatechartNestTest.cs
@@ -5,7 +5,7 @@ using NUnit.Framework;
 
 // ReSharper disable CheckNamespace
 
-namespace GameLoversEditor.StateChart.Tests
+namespace GameLoversEditor.Statechart.Tests
 {
 	[TestFixture]
 	public class StatechartNestTest
@@ -36,7 +36,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void BasicSetup()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimple, true,false));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimple, true,false));
 
 			statechart.Run();
 
@@ -55,7 +55,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void BasicSetup_WithoutTarget()
 		{
-			var statechart = new Statechart(factory => SetupNest_WithoutTarget(factory, _event2, SetupSimple));
+			var statechart = new StateMachine(factory => SetupNest_WithoutTarget(factory, _event2, SetupSimple));
 
 			statechart.Run();
 			statechart.Trigger(_event2);
@@ -75,7 +75,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void InnerEventTrigger()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,false));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,false));
 
 			statechart.Run();
 			statechart.Trigger(_event1);
@@ -98,7 +98,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void InnerEventTrigger_ExecuteFinal_SameResult()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,false));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,false));
 
 			statechart.Run();
 			statechart.Trigger(_event1);
@@ -121,7 +121,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void InnerEventTrigger_NotExecuteExit_SameResult()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimpleEventState, false,false));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimpleEventState, false,false));
 
 			statechart.Run();
 			statechart.Trigger(_event1);
@@ -144,7 +144,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void OuterEventTrigger()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,false));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,false));
 
 			statechart.Run();
 			statechart.Trigger(_event2);
@@ -167,7 +167,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void OuterEventTrigger_ExecuteFinal()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,true));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,true));
 
 			statechart.Run();
 			statechart.Trigger(_event2);
@@ -190,7 +190,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void OuterEventTrigger_NotExecuteExit()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimpleEventState, false,false));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimpleEventState, false,false));
 
 			statechart.Run();
 			statechart.Trigger(_event2);
@@ -213,7 +213,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void OuterEventTrigger_NotExecuteExit_ExecuteFinal()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimpleEventState, false,true));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimpleEventState, false,true));
 
 			statechart.Run();
 			statechart.Trigger(_event2);
@@ -236,7 +236,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void NestedStates_InnerEventTrigger()
 		{
-			var statechart = new Statechart(SetupLayer0);
+			var statechart = new StateMachine(SetupLayer0);
 
 			statechart.Run();
 			statechart.Trigger(_event1);
@@ -269,7 +269,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void NestedStates_OuterEventLayer0()
 		{
-			var statechart = new Statechart(SetupLayer0);
+			var statechart = new StateMachine(SetupLayer0);
 
 			statechart.Run();
 			statechart.Trigger(_event2);
@@ -302,7 +302,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void InnerEventTrigger_RunResetRun()
 		{
-			var statechart = new Statechart(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,false));
+			var statechart = new StateMachine(factory => SetupNest(factory, _event2, SetupSimpleEventState, true,false));
 
 			statechart.Run();
 			statechart.Trigger(_event1);
@@ -328,7 +328,7 @@ namespace GameLoversEditor.StateChart.Tests
 		[Test]
 		public void StateTransitionsLoop_ThrowsException()
 		{
-			Assert.Throws<InvalidOperationException>(() => new Statechart(factory =>
+			Assert.Throws<InvalidOperationException>(() => new StateMachine(factory =>
 			{
 				var initial = factory.Initial("Initial");
 				var nest = factory.Nest("Nest");